Make sql objects multithreaded friendly
Really quick hack to make the SQL database access and objects more multithread friendly. For most cases, no one will care but if the bot is encapsulated into a thread, this will make it such that someone could recreate the SQL environment on run.

Appears to work properly in tests. Alternatives could be to use scoped_sessions, so long as the engine is properly set up.

The idea here allows someone to destroy the previous database system on the bot's class initialization, and recreate it during the bot's run.