Update emote.py
If the twitch emotes api returns any error, this causes the entire bot to stop processing. Instead of allowing this to happen, we do a check to see if the 'emotes' key exists in the output json. If it does not, then we give up on trying to update the global emotes.